ADDED: Float Maximum Number

+++ src/com/generic/FindMaxNumber.java
@@ -1,13 +1,7 @@
 package com.generic;
 
 public class FindMaxNumber {
-
-	//public static void main(String[] args) {
-		//Integer[] i1= {1,0,0};
-        //int max= maxNumber(i1);
-        //System.out.println("our maximum number is"+" "+max);
-	//}
-	public static<E extends Comparable<E>>int maxNumber(E[] number)
+	public static<E extends Comparable<E>>float maxNumber(E[] number)
 	{
 	  E max=number[0];
 		for(E num:number)
@@ -16,7 +10,7 @@ public class FindMaxNumber {
 				max=num;
 		}
 		System.out.println(max);
-		return (int)max;
+		return (float)max;
 	}
 
 }
